Refactor and move autocomplete package to pkg/repo
Relocate the autocomplete package from pkg/core/edit to pkg/repo and adjust package declarations accordingly. Introduce AddWords method to enhance Dictionary functionality by allowing incremental additions of unique, sorted words.
