+++ pkg/repo/autocomplete_test.go
@@ -1,4 +1,4 @@
-package edit
+package repo
 
 import (
 	"sort"
